requirements.txt
django==3.0.11
djangorestframework-simplejwt
djangorestframework==3.11.0  # via -r requirements.in
flake8
isort
django-environ
